version pour maeva en cours de refactor

+++ lib/util/alert_helper.dart
@@ -1,33 +1,38 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:fluttertoast/fluttertoast.dart';
+import 'package:memories/models/post.dart';
+import 'package:memories/util/api_post_helper.dart';
+import 'package:memories/util/api_user_helper.dart';
 import 'package:memories/util/fire_helper.dart';
 import 'package:memories/view/my_material.dart';
 import 'package:launch_review/launch_review.dart';
 
+/// Classe permettant d'ouvrir des boites de dialogues afin d'informer l'utilisateur
 
 class AlertHelper {
-  //ALERT ERROR
+  ///ALERT ERROR
   Future<void> error(
       BuildContext context, String errorTitle, String error) async {
     Text title = Text(errorTitle);
     Text subtitle = Text(error);
     return showDialog(
-        context: context,
-        barrierDismissible: true,
-        builder: (BuildContext ctx) {
-          return (Theme.of(context).platform == TargetPlatform.iOS)
-              ? CupertinoAlertDialog(
-                  title: title,
-                  content: subtitle,
-                  actions: <Widget>[close(ctx, "OK")])
-              : AlertDialog(
-                  title: title,
-                  content: subtitle,
-                  actions: <Widget>[close(ctx, "OK")]);
-        });
+      context: context,
+      barrierDismissible: true,
+      builder: (BuildContext ctx) {
+        return (Theme.of(context).platform == TargetPlatform.iOS)
+            ? CupertinoAlertDialog(
+                title: title,
+                content: subtitle,
+                actions: <Widget>[close(ctx, "OK")])
+            : AlertDialog(
+                title: title,
+                content: subtitle,
+                actions: <Widget>[close(ctx, "OK")]);
+      },
+    );
   }
 
+  /// bouton Close Alert error
   FlatButton close(BuildContext ctx, String text) {
     return FlatButton(
       onPressed: (() => Navigator.pop(ctx)),
@@ -37,9 +42,8 @@ class AlertHelper {
     );
   }
 
-  //ALERTE DE SUPPRESSION DE PUBLICATION
-
-  Future<void> delete(BuildContext context, String id, String idFile,
+  ///ALERTE DE SUPPRESSION DE PUBLICATION
+  Future<void> delete(BuildContext context, ValueNotifier<List<Post>> notifier, String id, String idFile,
       String titleDialog, String subtitleDialog) async {
     Text title = Text(titleDialog);
     Text subtitle = Text(subtitleDialog);
@@ -53,26 +57,25 @@ class AlertHelper {
                   content: subtitle,
                   actions: <Widget>[
                       close(ctx, "Annuler"),
-                      deleteBtn(ctx, "Supprimer", id, idFile)
+                      deleteBtn(ctx, "Supprimer", id, idFile,notifier)
                     ])
               : AlertDialog(title: title, content: subtitle, actions: <Widget>[
                   close(ctx, "Annuler"),
-                  deleteBtn(ctx, "Supprimer", id, idFile)
+                  deleteBtn(ctx, "Supprimer", id, idFile,notifier)
                 ]);
         });
   }
 
+  ///Bouton de validation de suppression
   FlatButton deleteBtn(
-      BuildContext ctx, String text, String id, String idFile) {
+      BuildContext ctx, String text, String id, String idFile,ValueNotifier<List<Post>> notifier) {
     return FlatButton(
-      onPressed: (() => {
-            FireHelper().storage_posts.child(me.uid).child(idFile).delete(),
-            FireHelper()
-                .fire_user
-                .document(me.uid)
-                .collection("posts")
-                .document(id)
-                .delete(),
+      onPressed: (() async => {
+            ApiPostHelper().deletePost(id),
+            notifyListener(notifier,id),
+            if (idFile != null && idFile != ""){
+                FireHelper().storage_posts.child(me.uid).child(idFile).delete(),
+              },
             Navigator.of(ctx).popUntil((route) => route.isFirst)
           }),
       child: Text(
@@ -81,8 +84,12 @@ class AlertHelper {
     );
   }
 
-  //ALERTE DE DECONNECTION
+  notifyListener( ValueNotifier<List<Post>> notifier, String id){
+    notifier.value.removeWhere((element) => element.id==id);
+    notifier.notifyListeners();
+  }
 
+  ///ALERTE DE DECONNECTION
   Future<void> logOut(
       BuildContext context, String titleDialog, String subtitleDialog) async {
     Text title = Text(titleDialog);
@@ -106,17 +113,17 @@ class AlertHelper {
         });
   }
 
+  ///Bouton validation de deconnection
   FlatButton logOutBtn(BuildContext ctx, String text) {
     return FlatButton(
-      onPressed: (() => {Navigator.pop(ctx), FireHelper().logOut()}),
+      onPressed: (() => {ApiUserHelper().logOut(ctx)}),
       child: Text(
         text,
       ),
     );
   }
 
-  //ALERTE DE PUBLICATION DOUBLON LE MEME JOUR
-
+  ///ALERTE DE PUBLICATION DOUBLON LE MEME JOUR
   Future<void> overwrite(BuildContext context, String titleDialog,
       String subtitleDialog, Map<String, dynamic> map, String date) async {
     Text title = Text(titleDialog);
@@ -140,11 +147,13 @@ class AlertHelper {
         });
   }
 
+  ///bouton de validation de réécriture du post du jour
+  //TODO
   FlatButton overwriteBtn(
       BuildContext ctx, String text, Map<String, dynamic> map, String date) {
     return FlatButton(
       onPressed: (() => {
-            FireHelper()
+            /*FireHelper()
                 .fire_user
                 .document(map[keyUid])
                 .collection("posts")
@@ -165,9 +174,8 @@ class AlertHelper {
                 gravity: ToastGravity.BOTTOM,
                 timeInSecForIosWeb: 1,
                 fontSize: 16.0
-            )),
-        Navigator.of(ctx).popUntil((route) => route.isFirst)
-
+            )),*/
+            Navigator.of(ctx).popUntil((route) => route.isFirst)
           }),
       child: Text(
         text,
@@ -175,42 +183,38 @@ class AlertHelper {
     );
   }
 
-  //ALERTE NOUVELLE VERSION APPLICATION
-
-  Future<void> newVersion(BuildContext context, String titleDialog,
-      String subtitleDialog) async {
+  ///ALERTE NOUVELLE VERSION APPLICATION
+  Future<void> newVersion(
+      BuildContext context, String titleDialog, String subtitleDialog) async {
     Text title = Text(titleDialog);
     Text subtitle = Text(subtitleDialog);
     return showDialog(
         context: context,
-        barrierDismissible: true,
+        barrierDismissible: false,
         builder: (BuildContext ctx) {
           return (Theme.of(context).platform == TargetPlatform.iOS)
               ? CupertinoAlertDialog(
-              title: title,
-              content: subtitle,
-              actions: <Widget>[
-                newVersionBtn(ctx, "Télécharger")
-              ])
-              : AlertDialog(title: title, content: subtitle, actions: <Widget>[
-            newVersionBtn(ctx, "Télécharger")
-          ]);
+                  title: title,
+                  content: subtitle,
+                  actions: <Widget>[newVersionBtn(ctx, "Télécharger")])
+              : AlertDialog(
+                  title: title,
+                  content: subtitle,
+                  actions: <Widget>[newVersionBtn(ctx, "Télécharger")]);
         });
   }
 
-  FlatButton newVersionBtn(
-      BuildContext ctx, String text) {
+  ///Bouton de redirection vers Playstore
+  FlatButton newVersionBtn(BuildContext ctx, String text) {
     return FlatButton(
       onPressed: (() => {
-        LaunchReview.launch(
-          androidAppId: "fr.mencelt.memories",
-
-        )
-      }),
+            LaunchReview.launch(
+              androidAppId: "fr.mencelt.memories",
+            )
+          }),
       child: Text(
         text,
       ),
     );
   }
-
 }
